% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forint.R
\name{forint}
\alias{forint}
\title{Format numbers as Hungarian forints}
\usage{
forint(x)
}
\arguments{
\item{x}{number}
}
\value{
string
}
\description{
Format numbers as Hungarian forints
}
\examples{
forint(42)
forint(10000000.213214)
}
