% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_conversion_rate.R
\name{get_conversion_rate}
\alias{get_conversion_rate}
\title{Format numbers as Hungarian forints}
\usage{
get_conversion_rate(retried = 0)
}
\arguments{
\item{retried}{count the number of tries}
}
\value{
number
}
\description{
Format numbers as Hungarian forints
}
\examples{
get_conversion_rate(retried = 2)
}
