% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_conversion_rate.R
\name{get_conversion_rate_month}
\alias{get_conversion_rate_month}
\title{Format numbers as Hungarian forints}
\usage{
get_conversion_rate_month(
  fromCUR,
  toCUR,
  start_date = Sys.Date() - 30,
  end_date = Sys.Date(),
  retried = 0
)
}
\arguments{
\item{fromCUR}{base currency that you want to convert}

\item{toCUR}{final currency}

\item{start_date}{date}

\item{end_date}{date}

\item{retried}{count the number of tries}
}
\value{
\code{data.table} object
}
\description{
Format numbers as Hungarian forints
}
\examples{
get_exchange_rates('EUR', 'USD', start_date = '2020-05-12', end_date = '2020-05-13')
}
