% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_conversion_rate.R
\name{get_exchange_rates}
\alias{get_exchange_rates}
\title{Returns the exchange rates from one currency to another from a start to an end-date}
\usage{
get_exchange_rates(
  fromCUR,
  toCUR,
  start_date = Sys.Date() - 30,
  end_date = Sys.Date(),
  retried = 0
)
}
\arguments{
\item{fromCUR}{base currency that you want to convert}

\item{toCUR}{final currency}

\item{start_date}{date}

\item{end_date}{date}

\item{retried}{count the number of tries}
}
\value{
\code{data.table} object
}
\description{
Returns the exchange rates from one currency to another from a start to an end-date
}
\examples{
get_exchange_rates('EUR', 'USD', start_date = '2020-05-12', end_date = '2020-05-13')
get_exchange_rates('EUR', 'HUF', start_date = '2020-05-12', end_date = '2020-05-13')
}
