% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_conversion_rate.R
\name{get_usdhuf}
\alias{get_usdhuf}
\title{Returns the current exchange rate from USD to HUF}
\usage{
get_usdhuf(retried = 0)
}
\arguments{
\item{retried}{count the number of tries}
}
\value{
number
}
\description{
Returns the current exchange rate from USD to HUF
}
\examples{
get_usdhuf(retried = 2)
}
